Moved CourseCard to /pages dir
