+++ rev-learn/src/index.js
@@ -2,7 +2,6 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './index.css';
 import App from './App';
-import UserCourseCatalog from './pages/UserCourseCatalog/UserCourseCatalog';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
